TST: Refractor tests.
