+++ test/tester.f90
@@ -1,9 +1,8 @@
 program tester
     use iso_fortran_env
     use testdrive, only : run_testsuite, new_testsuite, testsuite_type
-    use testsuite_constants, only : collect_suite_constants
+    use testsuite_constants, only : collect_suite_version, collect_suite_year, collect_suite_constants
     use testsuite_uncertainties, only : collect_suite_uncertainties
-    use testsuite_year, only : collect_suite_year
     implicit none
     type(testsuite_type), allocatable :: testsuites(:)
     character(len=*), parameter :: fmt = '("#", *(1x, a))'
@@ -11,7 +10,8 @@ program tester
 
     stat = 0
 
-    testsuites = [new_testsuite("Year", collect_suite_year),&
+    testsuites = [new_testsuite("Version", collect_suite_version),&
+                  new_testsuite("Year", collect_suite_year),&
                   new_testsuite("Constants", collect_suite_constants),&
                   new_testsuite("Uncertainties", collect_suite_uncertainties)]
